--- modules/data-streams/qa/multi-node/build.gradle
@@ -1,28 +1,0 @@
-import org.elasticsearch.gradle.internal.info.BuildParams
-
-apply plugin: 'elasticsearch.internal-java-rest-test'
-
-File repoDir = file("$buildDir/testclusters/repo")
-
-tasks.named("javaRestTest").configure {
-  /* To support taking index snapshots, we have to set path.repo setting */
-  systemProperty 'tests.path.repo', repoDir
-}
-
-testClusters.matching { it.name == "javaRestTest" }.configureEach {
-  testDistribution = 'DEFAULT'
-  numberOfNodes = 4
-
-  setting 'path.repo', repoDir.absolutePath
-  setting 'xpack.security.enabled', 'false'
-  setting 'xpack.watcher.enabled', 'false'
-  setting 'xpack.ml.enabled', 'false'
-  setting 'xpack.license.self_generated.type', 'trial'
-  //disabling ILM history as it disturbs testDSXpackUsage test
-  setting 'indices.lifecycle.history_index_enabled', 'false'
-}
-
-if (BuildParams.inFipsJvm){
-  // Test clusters run with security disabled
-  tasks.named("javaRestTest").configure{enabled = false }
-}
