+++ modules/data-streams/src/javaRestTest/java/org/elasticsearch/xpack/datastreams/DataStreamsRestIT.java
