+++ modules/data-streams/src/yamlRestTest/resources/rest-api-spec/test/data_stream/10_basic.yml
