+++ modules/data-streams/src/yamlRestTest/resources/rest-api-spec/test/data_stream/50_delete_backing_indices.yml
