+++ modules/data-streams/src/yamlRestTest/resources/rest-api-spec/test/data_stream/60_get_backing_indices.yml
