+++ rest-api-spec/build.gradle
@@ -35,7 +35,6 @@ artifacts {
 
 testClusters.configureEach {
   module ':modules:mapper-extras'
-  module ':modules:data-streams'
   if (BuildParams.isSnapshotBuild() == false) {
     systemProperty 'es.index_mode_feature_flag_registered', 'true'
     systemProperty 'es.random_sampler_feature_flag_registered', 'true'
